python_track/personal_project/file_reader_stuff/requirements.txt
# LIBRARIES FOR FILE READER PROGRAM
pypdf==4.1.0
python-docx==1.1.0
pyttsx3==2.90